+++ carla_ros_bridge/src/carla_ros_bridge/traffic_participant.py
@@ -54,7 +54,7 @@ class TrafficParticipant(Actor):
         :return:
         """
         self.classification_age += 1
-        self.publish_transform(self.get_ros_transform())
+        self.publish_transform(self.get_ros_transform(None, None, str(self.get_id())))
         self.publish_marker()
         self.send_odometry()
 
@@ -68,7 +68,7 @@ class TrafficParticipant(Actor):
         odometry = Odometry(header=self.get_msg_header("map"))
         odometry.child_frame_id = self.get_prefix()
         odometry.pose.pose = self.get_current_ros_pose()
-        odometry.twist.twist = self.get_current_ros_twist()
+        odometry.twist.twist = self.get_current_ros_twist_rotated()
         self.publish_message(self.get_topic_prefix() + "/odometry", odometry)
 
     def get_object_info(self):