+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Latest changed
 
+*   Rework tf frame names
+*   ObjectSensor: Fix object twist
+*   Support loading OpenDRIVE map
 *   Traffic Lights: Only publish to /carla/traffic_lights on change
 *   Traffic Lights: Publish /carla/traffic_lights_info, containing the location and the trigger volume
+*   Added ROS Parameter to set the CARLA client timeout value for all nodes consistently
+
 ## CARLA-ROS-Bridge 0.9.8
 
 *   change Lidar range in meters